chore: rename batch of C files to C++
