+++ commands/uniq.cpp
@@ -1,3 +1,9 @@
+/*<<< WORK-IN-PROGRESS MODERNIZATION HEADER
+  This repository is a work in progress to reproduce the
+  original MINIX simplicity on modern 32-bit and 64-bit
+  ARM and x86/x86_64 hardware using C++23.
+>>>*/
+
 /* uniq - compact repeated lines		Author: John Woods */
 /*
  * uniq [-udc] [-n] [+n] [infile [outfile]]
